- Optimize numeric types of Object Model (replace some int by short or byte) - Update some OM providers in order to be compatible with the new OM definitions - Add some utils : InMemoryIdGen trait and median function

+++ src/main/scala/fr/proline/core/om/msi/MsQueryClasses.scala
@@ -9,7 +9,7 @@ package MsQueryClasses {
                  val id: Int,
                  val initialId: Int,
                  val moz: Double,
-                 val charge: Int,
+                 val charge: Byte,
                  
                  // Mutable optional fields
                  var properties: HashMap[String, Any] = new collection.mutable.HashMap[String, Any]
@@ -24,7 +24,7 @@ package MsQueryClasses {
                  override val id: Int,
                  override val initialId: Int,
                  override val moz: Double,
-                 override val charge: Int,
+                 override val charge: Byte,
                  val spectrumTitle: String,
                                   
                  // Mutable optional fields
